feat: implemented Breeding laboratory page
